Fixed update, delete, list operation

+++ mesh-worker-service/src/main/java/io/functionmesh/compute/rest/api/MeshComponentImpl.java
@@ -18,11 +18,13 @@
  */
 package io.functionmesh.compute.rest.api;
 
+import com.google.common.collect.Maps;
 import io.functionmesh.compute.functions.models.V1alpha1Function;
 import io.functionmesh.compute.functions.models.V1alpha1FunctionList;
 import io.functionmesh.compute.MeshWorkerService;
 import io.functionmesh.compute.sinks.models.V1alpha1Sink;
 import io.functionmesh.compute.sources.models.V1alpha1Source;
+import io.functionmesh.compute.util.CommonUtil;
 import io.functionmesh.compute.util.KubernetesUtils;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.Call;
@@ -49,6 +51,7 @@ import java.io.InputStream;
 import java.net.URI;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.function.Supplier;
 
@@ -109,28 +112,21 @@ public abstract class MeshComponentImpl implements Component<MeshWorkerService> 
                 clientAuthenticationDataHttps,
                 ComponentTypeUtils.toString(componentType));
         try {
-            Call componentCall =
-                    worker().getCustomObjectsApi()
-                            .getNamespacedCustomObjectCall(
-                                    group, version, KubernetesUtils.getNamespace(worker().getFactoryConfig()),
-                                    plural, componentName, null);
-            String clusterName;
-            if ("Source".equals(kind)) {
-                V1alpha1Source v1alpha1Source = executeCall(componentCall, V1alpha1Source.class);
-                clusterName = v1alpha1Source.getSpec().getClusterName();
-            } else if ("Sink".equals(kind)) {
-                V1alpha1Sink v1alpha1Sink = executeCall(componentCall, V1alpha1Sink.class);
-                clusterName = v1alpha1Sink.getSpec().getClusterName();
-            } else {
-                V1alpha1Function v1alpha1Function = executeCall(componentCall, V1alpha1Function.class);
-                clusterName = v1alpha1Function.getSpec().getClusterName();
+            String clusterName = null;
+            if (worker().getFactoryConfig() != null && worker().getFactoryConfig().getCustomLabels() != null) {
+                Map<String, String> customLabels = worker().getFactoryConfig().getCustomLabels();
+                clusterName = customLabels.get(CLUSTER_LABEL_CLAIM);
             }
+            if (clusterName == null) {
+                throw new RestException(javax.ws.rs.core.Response.Status.BAD_REQUEST, "Please set cluster name");
+            }
+            String hashName = CommonUtil.createObjectName(clusterName, tenant, namespace, componentName);
             Call deleteObjectCall = worker().getCustomObjectsApi().deleteNamespacedCustomObjectCall(
                     group,
                     version,
                     KubernetesUtils.getNamespace(worker().getFactoryConfig()),
                     plural,
-                    componentName,
+                    hashName,
                     null,
                     null,
                     null,
@@ -317,8 +313,22 @@ public abstract class MeshComponentImpl implements Component<MeshWorkerService> 
                                       final AuthenticationDataSource clientAuthenticationDataHttps) {
         List<String> result = new LinkedList<>();
         try {
-            String labelSelector = String.format("%s=%s,%s=%s",
-                    TENANT_LABEL_CLAIM, tenant, NAMESPACE_LABEL_CLAIM, namespace);
+            String labelSelector;
+            if (worker().getFactoryConfig() != null && worker().getFactoryConfig().getCustomLabels() != null) {
+                Map<String, String> customLabels = worker().getFactoryConfig().getCustomLabels();
+                String cluster = customLabels.get(CLUSTER_LABEL_CLAIM);
+                labelSelector = String.format(
+                        "%s=%s,%s=%s,%s=%s",
+                        CLUSTER_LABEL_CLAIM, cluster,
+                        TENANT_LABEL_CLAIM, tenant,
+                        NAMESPACE_LABEL_CLAIM, namespace);
+            } else {
+                labelSelector = String.format(
+                        "%s=%s,%s=%s",
+                        TENANT_LABEL_CLAIM, tenant,
+                        NAMESPACE_LABEL_CLAIM, namespace);
+            }
+
             Call call = worker().getCustomObjectsApi().listNamespacedCustomObjectCall(
                     group,
                     version,
@@ -334,7 +344,7 @@ public abstract class MeshComponentImpl implements Component<MeshWorkerService> 
                     null);
 
             V1alpha1FunctionList list = executeCall(call, V1alpha1FunctionList.class);
-            list.getItems().forEach(n -> result.add(n.getMetadata().getName()));
+            list.getItems().forEach(n -> result.add(n.getMetadata().getLabels().get(COMPONENT_LABEL_CLAIM)));
         } catch (Exception e) {
             log.error("failed to fetch functions list from namespace {}, error message: {}", namespace, e.getMessage());
         }
